- Change consommation Prevu to Long - Fix Batch_faker

+++ src/main/java/me/kadarh/mecaworks/domain/bons/BonEngin.java
@@ -70,7 +70,7 @@ public class BonEngin extends AbstractDomain {
     private Float consommationKm = 0f;
     private Float consommationH = 0f;
     private Long nbrHeures = 0L;
-    private Float consommationPrevu = 0f;
+    private Long consommationPrevu = 0L;
     private Long chargeHoraire = 0L;
